moved prereq tree to components
