+++ frontend/src/components/PrerequisiteTree/config.js
