+++ frontend/src/components/PrerequisiteTree/utils.js
