+++ frontend/src/pages/CourseSelector/CourseDescription/CourseDescription.jsx
@@ -11,7 +11,7 @@ import axiosRequest from "config/axios";
 import { inDev, TERM, TIMETABLE_API_URL } from "config/constants";
 import { setCourse } from "reducers/coursesSlice";
 import prepareUserPayload from "../utils";
-import PrerequisiteTree from "./PrerequisiteTree";
+import PrerequisiteTree from "components/PrerequisiteTree";
 import LoadingSkeleton from "./LoadingSkeleton";
 import PlannerButton from "./PlannerButton";
 import S from "./styles";
